packages/agents/bidi-agent/pyproject.toml
[project]
name = "idp_v2.bidi-agent"
version = "0.1.0"
description = "Bidi Agent"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "asyncio>=4.0.0",
    "strands-agents[bidi-all]>=1.22.0",
]
